OK, pixel R G B values are now int

+++ src/ledMatrix.py
@@ -67,7 +67,7 @@ def getLineImage(centre, angle, length, colour, width=8, height=8):
     colour = np.array(colour)
     ledValues = np.array([ledIntensities, ledIntensities, ledIntensities])
     ledValues = ledValues.T*colour
-    #ledValues = ledValues
+    ledValues = ledValues.astype(int)
     return ledValues
